Correct and refactor address meta loading logic
The existing logics doesn't handle the case, when both the billing and shipping address have to be checked via web api and both
have minor correction. In this case only the billing address would be corrected and the shipping retains the old address meta leading to
the display of the correction modal. Which is faulty.
So this change extracts the autocorrection application logic into a separate method, so that it can be called both for
address meta coming from web api, but also for the address meta coming from cache.

Ref: J5P-45