Dont load meta to html if delivery template is selected
Sometimes the address meta are loaded to the page, when there is  a form with address, but its invisible, such as
in the checkout in jtl 5.2 when multiple delivery address templates are available.
In order to circumvent this problem two more checks are added to replace old logic.
1. Check if the form with shipping address contains any data.
2. Check if there is no delivery template checkbox selected
If both are true ,then its assumed safe to load the address meta to shipping address.

Fixes: J5P-47