Remove early smarty initiation
In order to process some information we were initiating smarty in the bootstrap
of the plugin. Smarty object was used by metahandler to check, if any delivery
address templates were preselected in html.
This however caused a problem with jtl search plugin, because jtl search plugin
uses SMARTY_INC hook to add some data to the page. Basically it assumes that any Smarty
object will be initiated after all the plugin hooks were registered, which was not the case
with our plugin.
While from my perspective its a questionable decision from jtl software team, especially
given that they have HOOK_SMARTY_OUTPUTFILTER that would also work well, i changed the
was we get Smarty object in the metahandler to facade accessor. Its less stable, and not
testable, but well...

Ref: J5P-52