Add an external address meta invalidation to address check results
In some case the address form is closed while the address check is triggered. When results
arrive, the form that can receive a correction is not valid anymore.
So with this modification, we add another check to invalidate results of the addresscheck,
which in turn prevents any reactions like modal appearing.
The result is invalidated only for delivery address when in checkout a deliveryaddress is selected,
which closes the previosly opened form for shipping address while also triggering the address check.

Ref: J5P-50