Reset address status for shipping address when selecting template
When the user is selecting the delivery template in jtl 5.2 checkout, it closes the form with shipping address.
If the AMS object was bound to it though, it doesnt get updated properly, essentially leaving the status codes
in CSS Classes. So the form is empty, btu has color codes for input fields.
With this change the address status gets reset, when user selects a delivery address template.

Ref: J5P-50