Load meta only for Lieferadresse on checkout page
In the previos logic we checked if the actual delivery template was selected and if yes,
then for this delivery template the address meta was loaded.
It turnes out its possible, that the data shows that a delivery template is selected,
even though its actually not -> the id of the delivery template is replaced with an
ID of a delivery address from previous orders.
When tryging to load delivery address template with this wrong ID
we would get an empty address, which leads to a display of wrong correction modal.
So to fix it, we load from Lieferadress Object instead. This should contain either a copy
of previously selected delivery address template or something else.

Ref: J5P-56