Remove internal ID check from order comment generation
In the current implementation the plugin checks if Lieferadresse has an ID (kLeferadresse)
or the Rechnungsadresse has an Id, before proceeding with comment generation.
This doesn't work in some cases, when the datas have no ID at all or when the Lieferaddresse
object has no ID, and Rechnungsadresse has no statuses.
So in order to address this issue we remove the checks for ID's. we assume that in the SESSION
the correct statuscodes are loaded or there are no statuscodes.
Risk is: we might need to invalidate some statuscodes for some strange edgecases.

Ref: J5P-57