Fix wrong status mapping
When the automatic correction is applied, there is a need to fetch new statuscodes to describe the new address. In order to save on an request to the API we just generate them. Because the corrected variant is expected to always be correct, we just set all the field statuses to correct and set the general statuses. As the address object might contain different fields, we have a mapping table that maps field names to corresponding statuses.
However the mapping table had the same value for all fields, which is wrong. This commit fixes this.

Ref: J5P-73