Block ams logic for some payment method in checkout last page
Some paymentd methods preauthorize the address before the last page in checkout. However the current implementation of enderecos address validation and correction uses the last page as a last resort to notify the customer about problems in his address and allows him to use proposed address correction from modal (in gets copied and saved in DB and SESSION automatically) or even have the address
corrected automatically behind the scenes.
Now if the address is corrected in this last step, it might cause an authirization error with some payment provides such as klarna payment. The address essentially is considered manipulated and fraudy and the order can not be closed.
Long term solution would be to extend the logic of how the corrections are done by the endereco plugin. Short term would be to disable
the correction logic in the checkout for those kind of payment methods.
This commit adds the disabling logic.

Ref: EST-414