Add null coalescing operator for http_referrer
When sending queries to Endereco API, one of the important pieces of information
is where the query is coming from. For server-side requests, we use
$_SERVER['HTTP_REFERER'] for this information, which can sometimes be null,
generating a PHP warning. This warning might be visible in the frontend,
implying low quality of our plugin.

Fix the issue by using a null coalescing operator and providing a default
fallback value in case of null.

Ref: J5P-76