Remove button from reset script
In some platforms like JTL the submit button get disabled when a form is being submitted. Its possible that endereco plugin
interrupts the submit to validate the address. So it can happen, that the form will be locked, as the submit got interrupted and submit button
disabled.
To prevent this in endereco plugin there is a script that resets the submit button. This script has a CSS selector for the button.
Recently we found out, that this selector is not cevering the case, where the buton is an <input> element.
So the selector needs to be a bit more general.

Ref: EST-448