Add compatibility with 5.4
Since 5.2 the direct access to io.php has been marked as deprecated. Since 5.4 the file io.php has been remoevd.
Which broke out AJAX logic and made it impossible for a user to confirm/select address in checkout via endereco modal.
In the fix we add legacy support for 5.1 and change the way the io is accessed (through router)

Ref: J5P-80