Fix floating label issue in JTL 5.5.0+ by removing faulty css class
Additionally to the existing typeahead removal of CSS classes on
the input element, the "typeahead-required" class from the
wrapping element must also be removed.

Without this, it causes a visual bug in the user registration
form where the city input label doesn't move out of the input
field when focused, blocking the user's view and preventing
proper form interaction.

Fixes #3