Update js-sdk to version 1.13.0 and adapt init function
This commit updates the @endereco/js-sdk dependency to version 1.13.0.

It modifies the address validation module initialization to support
asynchronous initialization, ensuring compatibility with the updated library.

Without using async the initialization throws errors in the console because EAO.util is not ready in the moment the initialization runs.

A few function parameters were renamed to better reflect what they stand for, ensuring consistency to the documentation.

Fixes #13