Add option to disable order attribute creation
Adds a configuration setting to disable saving address validation results to order attributes. This allows store owners to control whether address validation data is stored in the database.

The setting is accessible via the plugin configuration. The default value is off.

This changes the default behaviour of the plugin. Before this change every order created multiple entries in the order attribute table.

Fixes #18